pyproject.toml
[tool.poetry]
name = "poqrl"
version = "0.1.0"
description = ""
authors = ["micha <michael.blot@walnut.ai>"]

[tool.poetry.dependencies]
python = "3.10.4"
Pillow = "^9.1.1"
black = "^22.3.0"
pylint = "^2.13.9"
pytest = "^7.1.2"
plotly = "^5.8.0"
numpy = "^1.22.4"
tqdm = "^4.64.0"
tensorflow = "^2.9.1"
notebook = "^6.4.11"
typer = "^0.4.1"
jax = "^0.3.13"
pandas = "^1.5.3"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
